feat(arc-375): sync es results to state, create media store, get genre aggs
