+++ src/modules/media/types/index.ts
@@ -1,36 +1,30 @@
 import { MediaTypes } from '@shared/types';
 
-export interface MediaInfo {
-	schema_in_language: unknown | null;
-	dcterms_available: string;
-	schema_creator?: {
-		Archiefvormer?: string[];
-		productionCompany?: string[];
-		Maker?: string[];
-	};
-	schema_identifier: string;
-	schema_description?: string;
-	schema_publisher?: {
-		Publisher: string[];
-	};
-	schema_duration: string;
-	schema_abstract?: string;
-	premis_identifier: string;
-	schema_genre?: string;
-	schema_date_published?: string;
-	schema_license?: string[];
-	schema_date_created?: string;
-	schema_contributor: unknown | null;
-	schema_maintainer: {
-		schema_identifier: string;
-	}[];
-	dcterms_format: MediaTypes;
-	schema_name: string;
-	// TODO: See if this is still necessary once resolved in proxy
-	type?: string;
+export interface MediaContactInfo {
+	email: string;
+	telephone: string;
+	address: MediaAddress;
 }
 
-// Mapped intelectual entity object
+export interface MediaAddress {
+	street: string;
+	postalCode: string;
+	locality: string;
+	postOfficeBoxNumber: string;
+}
+
+export interface MediaRepresentation {
+	name: string;
+	alternateName: string;
+	description: string;
+	meemooFragmentId: string;
+	dctermsFormat: string;
+	transcript: string;
+	dateCreated: string;
+	id: string;
+	files: File[];
+}
+
 export interface Media {
 	id: string;
 	premisIdentifier: any;
@@ -73,31 +67,6 @@ export interface Media {
 	dateCreatedLowerBound: string;
 	ebucoreObjectType: string;
 	representations: MediaRepresentation[];
-}
-
-export interface MediaContactInfo {
-	email: string;
-	telephone: string;
-	address: MediaAddress;
-}
-
-export interface MediaAddress {
-	street: string;
-	postalCode: string;
-	locality: string;
-	postOfficeBoxNumber: string;
-}
-
-export interface MediaRepresentation {
-	name: string;
-	alternateName: string;
-	description: string;
-	meemooFragmentId: string;
-	dctermsFormat: string;
-	transcript: string;
-	dateCreated: string;
-	id: string;
-	files: File[];
 }
 
 export class MediaSearchFilters {
