+++ src/modules/shared/store/media/media.select.ts
@@ -1,0 +1,5 @@
+import { GetMedia } from '@shared/types';
+
+import { AppState } from '../store.types';
+
+export const selectMediaResults = (state: AppState): GetMedia | undefined => state.media.results;
