+++ src/modules/shared/store/media/media.types.ts
@@ -1,0 +1,5 @@
+import { GetMedia } from '@shared/types';
+
+export interface MediaState {
+	results?: GetMedia;
+}
